Hosts can Kick players by pressing on them
In Event Details, if you are the host, you can click on attendees to
decide whether or not you want to kick them. In this version kicking
forces the players to Leave the event, not "ban" them. No status changes
yet.

Host cannot kick themselves.

Modified LeaveEventAsAttendee to take in 2 parameters in order to reduce
code redundancy.