Finally got push to work (Invitations)
https://www.parse.com/docs/android/guide#push-notifications-troubleshooting

Had to turn on the setting for client pushes in parse.

"Please note that SDKs that use a Client Key, such as the Android SDK,
can only send push notifications if Client Push is enabled in your Parse
app's Push Notification settings."

Biggest fucking clue.

In Workaround.pushToRecipient, put in userid as a string and a text
message that will appear on their devices